Combine the library into a single module
Update the library so that it acts as a single module. Remove all of
the individual go.mod packages and make sure that the internal/errors
reference has a full github.com/njones/socketio path.

Also add in the build directives for the different protocol and server
combinations that can happen. Update to the latest passing tests.