Clean up the socketio packet protocol struct
Move the binary incoming/outgoing streams code to a seperate struct
so that we can embed them into all following protocols easily.

Clean up the test with indenting.

Push the type aliasing a bit deeper so that Server options can be
exchanged between the socket.io and engine.io servers as expected.