Code cleanup for tests, loops and io.EOF responses
Removed the "Receive" label from around for loops and just return
instead of breaking on a label. This should be a bit easier to
read and understand what is happening in the future.

Removed the binary test from `server.v1_test` because the version 1
of the socket.io server protocol. Binary data is introduced in version 2
of the socket.io protocol.

Propagated the io.EOF response from the Copy function when reading a
packet type within the `engineio/protocol`. So that io.EOF can be
captured by the encoding/decoding function that is writing/reading the
packet.

Simple test formatting for consistency (and assertion evaluation)