Fix EIO protocol Handshake and  HTTP Polling
Fix the HandshakeV3 protocol by changing it to a pointer so that changes
can be reflected when upadted and so that the type matches when checking
the packet data during the type conversion to the wire representation
(bytes).
Fix the polling transport so that it sends more than one packet in
a payload when there are several packets in the queue.