Add ReadBinary interface and small fixes
Add the ReadBinary method to satisfy a ReadBinary interface to allow SIO
packet version 4 and 5 to accept binary data from clients.
Update the namespace struct so that the Len() count is accurate for
empty namespaces.
Fix the msgpack logic so that it can accept non-base64 encoded data values.
Add a Reader that can propagate errors from io.Copy to the reader, so that
any errors can be properly managed even if it comes from the io.Copy
embedded within a go-routine.