Fix EIO Protocol Len() sum and EOH sessionIDs
Fix issues around calculating the length of the pingTimeout and
pingInterval values (as seconds).
Refactor the PayloadV3 to remove duplicate code, and push the processing
to PayloadV2 as expected.
Update the EOH error to include the current SessionID so that it can be
used to populate the SIO transport before it's called the second time.
This allows for SIO v2 auto connections during the Handshake phase.