Refactor Protocol/Payload - decoding/encoding
Updates the EngineIO protocol and payload encoding/decoding methods to
be more clear. Updates the Binary support to not be based on if the
packet is Binary but rather if the protocol supports binary.
Updates the JSON and Base64 encoding to satisfy a common interface and
be added to the rw.Writer and rw.Reader structs so they can be used in
a fluid manner.