Update functional options (#42)
Refactor most functional options to use the same underlining interface
which will help with inter-op and passing them along. For instance, now
a transport functional option can be passed in during the SocketIO
server setup. Also, simplified the execution path for applying options
previously it was a O(n^n) application process of options. This has been
simplified to O(n*m), and gives a much more logical execution path of
when an option will be applied.