Refactor tests to allow hooks for advanced testing
Update the tests for the socketio server, so that we can start to add hooks around asserts, then we can test errors and make other advances tests around the happy path, and non-happy path.
The commit doesn't add the hooks, but sets up the tests so that we can add a variable number of parameters to test, which in the future will include hooks.