Fix the SocketIO connect protocol (#65)
* Fix the SocketIO connect protocol
* Update SocketIO server and packet tests

When SocketIO connects to a namespace the client expects a return connect call. That was only happening during auto-connect, and not during normal server connections. Now the code correctly returns a connection to the client for every successful server connection.

Updates the tests so they reflect the change in how namespaces are processed. The packet tests are more flexible (like the SocketIO server tests). Added an extra check around polling in the server tests, so they aren't flaky.

[fixes #61]