Update the callback.Wrap struct to work with maps
The callback wrap method didn't understand the map type when doing reflection. This commit adds the map type and serializes to JSON so that the Unserialize() on the map object will work as expected.

NOTE: that custom Serialize() and Unserialize() objects can be always used instead of the generic callback.Wrap struct. Examples were added to the callback_test.go file.