Add a socketio.WithPath to change the SIO path
The engineio.WithPath option would change the EngineIO path, but left the SocketIO path untouched which caused issues with the service because the SocketIO and EngineIO internal paths didn't match.

This commit adds a socketio.WithPath option to change SocketIO paths (and the underlining EngineIO path).

This commit also abstracts the EngineIO WithPath option to guarantee that the code will compile even if the option is removed when building a low version copy of the library. In this case it's not really important, but this is to showcase how it will work for new and refactored options.

Tests around adding the socketio.WithPath have been added separately outside of the integration tests, because they are testings a specific server configuration rather than a SocketIO implementation.